Split out MixedVecSpec integration tests
This will speed up the typical test flow by separating the execution
driven tests out.

Also rename the file containing VecIntegrationSpec to match the class
name.
