+++ integration-tests/src/test/scala/chiselTest/VecIntegrationSpec.scala
